stateful clients + tamuna without compression

+++ baselines/tamuna/tamuna/client.py
@@ -6,6 +6,8 @@ from typing import Callable, Dict, List, Tuple
 import os
 import flwr as fl
 import copy
+
+import numpy as np
 import torch
 from flwr.common.typing import NDArrays, Scalar
 from hydra.utils import instantiate
@@ -13,6 +15,10 @@ from omegaconf import DictConfig
 from torch.utils.data import DataLoader
 
 from tamuna.models import test, train
+
+
+def apply_compression(net, mask):
+    return net
 
 
 class FlowerClient(fl.client.NumPyClient):
@@ -30,19 +36,24 @@ class FlowerClient(fl.client.NumPyClient):
         self.trainloader = trainloader
         self.device = device
         self.learning_rate = learning_rate
-        self.control_variate = self.__model_zeroed_out()
+        self.control_variate = None
         self.old_compression_mask = None
         self.old_compressed_net = None
         self.cid = cid
 
-        with open(f"{self.cid}_state.bin", "wb") as f:
-            state = (
-                self.control_variate,
-                self.old_compression_mask,
-                self.old_compressed_net,
-            )
-            pickle.dump(state, f, protocol=pickle.HIGHEST_PROTOCOL)
-            print(f"Client {self.cid}.")
+        state_file_name = f"{self.cid}_state.bin"
+
+        if not os.path.exists(state_file_name):
+            with open(state_file_name, "wb") as f:
+                state = (
+                    self.control_variate,
+                    self.old_compression_mask,
+                    self.old_compressed_net,
+                )
+                pickle.dump(state, f, protocol=pickle.HIGHEST_PROTOCOL)
+                print(f"Client {self.cid} state created.")
+        else:
+            print(f"Client {self.cid} state already exists.")
 
     def get_parameters(self, config: Dict[str, Scalar]) -> NDArrays:
         """Returns the parameters of the current net."""
@@ -54,14 +65,6 @@ class FlowerClient(fl.client.NumPyClient):
         state_dict = OrderedDict({k: torch.Tensor(v) for k, v in params_dict})
         self.net.load_state_dict(state_dict, strict=True)
 
-    def __model_zeroed_out(self):
-        control_variate = copy.deepcopy(self.net)
-        state_dict = OrderedDict(
-            {k: torch.zeros_like(v) for k, v in self.net.state_dict().items()}
-        )
-        control_variate.load_state_dict(state_dict, strict=True)
-        return control_variate
-
     def fit(
         self, parameters: NDArrays, config: Dict[str, Scalar]
     ) -> Tuple[NDArrays, int, Dict[str, int]]:
@@ -69,6 +72,7 @@ class FlowerClient(fl.client.NumPyClient):
         self.set_parameters(parameters)
 
         # load compression mask from file {self.cid}_mask.bin
+        mask = np.zeros(1)  # todo
 
         with open(f"{self.cid}_state.bin", "rb") as f:
             state = pickle.load(f)
@@ -78,14 +82,26 @@ class FlowerClient(fl.client.NumPyClient):
                 self.old_compressed_net,
             ) = state
 
-        train(
+        self.net, self.control_variate = train(
             self.net,
             self.trainloader,
             self.device,
             epochs=config["epochs"],
+            eta=config["eta"],
             control_variate=self.control_variate,
-            learning_rate=self.learning_rate,
+            server_net=copy.deepcopy(self.net),
+            lr=self.learning_rate,
         )
+
+        self.net = apply_compression(self.net, mask)
+
+        with open(f"{self.cid}_state.bin", "wb") as f:
+            state = (
+                self.control_variate,
+                mask,
+                self.net,
+            )
+            pickle.dump(state, f, protocol=pickle.HIGHEST_PROTOCOL)
 
         return self.get_parameters({}), len(self.trainloader), {}
 
@@ -115,6 +131,8 @@ def gen_client_fn(
 
     def client_fn(cid: str) -> FlowerClient:
         """Create a Flower client representing a single organization."""
+
+        print(f"Generating client {cid}.")
 
         # Load model
         device = torch.device(device=client_device)
